% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{parseDatetime}
\alias{parseDatetime}
\title{Parse a Datetime object from a string}
\usage{
parseDatetime(txt, fmt = "\%Y-\%m-\%dT\%H:\%M:\%E*S\%Ez", tzstr = "UTC")
}
\arguments{
\item{txt}{A string from which a datetime is to be parsed}

\item{fmt}{A string with the format, which is based on \code{strftime} with some
extensions; see the CCTZ documentation for details.}

\item{tzstr}{The local timezone for the desired format}
}
\value{
A Datetime object
}
\description{
Parse a Datetime object
}
\details{
An alternative to \code{as.POSIXct} based on the CCTZ library
}
\examples{
ds <- getOption("digits.secs")
options(digits.secs=6) # max value
parseDatetime("2016-12-07 10:11:12",        "\%Y-\%m-\%d \%H:\%M:\%S");   # full seconds
parseDatetime("2016-12-07 10:11:12.123456", "\%Y-\%m-\%d \%H:\%M:\%E*S"); # fractional seconds
parseDatetime("2016-12-07T10:11:12.123456-00:00")  ## default RFC3339 format
options(digits.secs=ds)
}
\author{
Dirk Eddelbuettel
}

